add some cool animations when switching tabs
