+++ lib/components/dm-helper/panels/DMHelperTabPanel.tsx
@@ -1,0 +1,15 @@
+import { TabPanel, TabPanelProps } from '@chakra-ui/react';
+
+export default function DMHelperTabPanel({
+  index,
+  current,
+  children,
+  ...rest
+}: TabPanelProps & { index: number; current: number }) {
+  if (index !== current) return null;
+  return (
+    <TabPanel maxH="100%" h="100%" px="0" {...rest}>
+      {children}
+    </TabPanel>
+  );
+}
